refactor(back): divide responsabiliddades entre modulos
