--- backend/src/routes/server.routes.ts
@@ -1,8 +1,0 @@
-import { Router } from 'express'
-import { getServerStatus } from '../controllers/server.controller'
-
-const router = Router()
-
-router.get('/status', getServerStatus)
-
-export default router
